changed npmt-test from criterium to country
